feat(onboarding): remove onboarding tooltips
- remove onbarding tooltips from search
- add a modal based on the user state for the welcome message
- add logic to present this message once, based on local storage data
