+++ e2e/tutor.spec.ts
@@ -162,6 +162,8 @@ test.describe('Tutor', () => {
       }
     });
     await page.goto('/tutor');
+    // close welcome modal
+    await page.getByTestId('close').click();
   });
 
   test('should display input type file', async ({ page }) => {
