+++ src/stores/featureFlip.ts
@@ -3,8 +3,6 @@ type FeatureFlip = {
   [key: string]: boolean;
 };
 export const useFeatureFlipStore = defineStore('featureFlip', () => {
-  // const { VITE_ENVIRONMENT } = import.meta.env;
-  // const env = VITE_ENVIRONMENT as 'dev' | 'prod';
   const featureFlip: FeatureFlip = {
     tutor: true
   };
