+++ src/views/SearchSDG.vue
@@ -3,8 +3,7 @@ import Button from '@/components/ButtonComponent.vue';
 import TextArea from '@/components/TextAreaComponent.vue';
 import FullpageTemplate from '@/components/FullpageTemplate.vue';
 import { useSearchStore } from '@/stores/search';
-import { useUserStore } from '@/stores/user';
-import { onBeforeMount, watch } from 'vue';
+import { onBeforeMount } from 'vue';
 import SourcesListComponent from '@/components/SourcesListComponent.vue';
 import SearchIcon from '@/components/icons/SearchIcon.vue';
 import SelectSubject from '@/components/dropdowns/SubjectSelector.vue';
@@ -12,75 +11,22 @@ import SearchWrapper from '@/components/SearchWrapper.vue';
 import SelectSource from '@/components/dropdowns/SourcesSelector.vue';
 import SDGSelector from '@/components/dropdowns/SDGSelector.vue';
 import DeleteButton from '@/components/DeleteButton.vue';
-import OnboardingTooltip from '@/components/OnboardingTooltip.vue';
 import Pill from '@/components/PillComponent.vue';
-import ModalComponent from '@/components/ModalComponent.vue';
 import ErrorComponent from '@/components/ErrorComponent.vue';
 
 const store = useSearchStore();
-const user = useUserStore();
 
 onBeforeMount(() => {
   store.getCorpus();
 });
-
-watch(
-  () => store.searchInput,
-  (a, b) => {
-    if (a.length > b.length && user.onboardingSearch.step === 'INPUT') {
-      user.updateSearchState('INPUT');
-    }
-  }
-);
-watch(
-  () => store.hasFilters,
-  (a) => {
-    if (a && user.onboardingSearch.step === 'FILTERS') {
-      user.updateSearchState('FILTERS');
-    }
-  }
-);
-watch(
-  () => store.isFetchingSources,
-  (a) => {
-    if (a && user.onboardingSearch.step === 'SEND') {
-      user.updateSearchState('SEND');
-    }
-  }
-);
 </script>
 
 <template>
   <FullpageTemplate>
     <template #top>
-      <ModalComponent
-        v-if="
-          user.onboardingSearch.state === 'incomplete' && user.onboardingSearch.step === 'WELCOME'
-        "
-        :title="$t('onboarding.welcome.title')"
-        :message="$t('onboarding.welcome.description')"
-        :isOpen="true"
-        :onClose="() => user.updateSearchState('WELCOME', true)"
-      >
-        <template #actions>
-          <button class="button is-primary" @click="user.updateSearchState('WELCOME')">
-            {{ $t('next') }}
-          </button>
-        </template>
-      </ModalComponent>
       <ErrorComponent v-if="store.hasError" />
       <SearchWrapper>
         <template #textArea>
-          <OnboardingTooltip
-            arrowPosition="left"
-            v-if="
-              user.onboardingSearch.state === 'incomplete' && user.onboardingSearch.step === 'INPUT'
-            "
-            :description="$t('onboarding.search.INPUT')"
-            :onClose="() => user.updateSearchState('INPUT', true)"
-            :onNext="() => user.updateSearchState('INPUT')"
-            :buttonText="$t('next')"
-          />
           <div class="delete-button-wrapper" v-if="store.searchInput">
             <DeleteButton :action="store.$reset" :delText="$t('clearSearch')" />
           </div>
@@ -96,16 +42,6 @@ watch(
           </div>
         </template>
         <template #filters>
-          <OnboardingTooltip
-            v-if="
-              user.onboardingSearch.state === 'incomplete' &&
-              user.onboardingSearch.step === 'FILTERS'
-            "
-            :description="$t('onboarding.search.FILTERS')"
-            :onClose="() => user.updateSearchState('FILTERS', true)"
-            :buttonText="$t('next')"
-            :onNext="() => user.updateSearchState('FILTERS')"
-          />
           <SelectSource />
           <SDGSelector />
           <SelectSubject
@@ -116,16 +52,6 @@ watch(
           />
 
           <div class="is-relative search-button">
-            <OnboardingTooltip
-              v-if="
-                user.onboardingSearch.state === 'incomplete' &&
-                user.onboardingSearch.step === 'SEND'
-              "
-              :description="$t('onboarding.search.SEND')"
-              :onClose="() => user.updateSearchState('SEND', true)"
-              :buttonText="$t('next')"
-              :onNext="() => user.updateSearchState('SEND')"
-            />
             <Button
               class="search-button"
               :aria-label="$t('search')"
